feat: 기록 입력 페이지 퍼블 작업 진행
