+++ frontend/eodi-web/app/field-notes/new/_components/FieldDescription.jsx
@@ -1,0 +1,1 @@
+export defaul
